[CHORE] page merge & divide component
